+++ BraintreeCard/BTCardNonce.m
@@ -7,13 +7,18 @@
                   description:(NSString *)description
                   cardNetwork:(BTCardNetwork)cardNetwork
                       lastTwo:(NSString *)lastTwo
+                     lastFour:(NSString *)lastFour
                     isDefault:(BOOL)isDefault
-                     cardJSON:(nonnull BTJSON *)cardJSON
-              authInsightJSON:(nullable BTJSON *)authInsightJSON {
-    self = [super initWithNonce:nonce localizedDescription:description type:[BTCardNonce typeStringFromCardNetwork:cardNetwork] isDefault:isDefault];
+                     cardJSON:(BTJSON *)cardJSON
+              authInsightJSON:(BTJSON *)authInsightJSON {
+    self = [super initWithNonce:nonce
+           localizedDescription:description
+                           type:[BTCardNonce typeStringFromCardNetwork:cardNetwork]
+                      isDefault:isDefault];
     if (self) {
         _cardNetwork = cardNetwork;
         _lastTwo = lastTwo;
+        _lastFour = lastFour;
         _binData = [[BTBinData alloc] initWithJSON:cardJSON[@"binData"]];
         if ([cardJSON[@"details"][@"bin"] asString]) {
             _bin = [cardJSON[@"details"][@"bin"] asString];
@@ -91,17 +96,21 @@
         authInsightJson = cardJSON[@"authenticationInsight"];
     }
     
-    return [[[self class] alloc] initWithNonce:[cardJSON[@"nonce"] asString]
+    return [[self.class alloc] initWithNonce:[cardJSON[@"nonce"] asString]
                                    description:[cardJSON[@"description"] asString]
                                    cardNetwork:[self.class cardNetworkFromGatewayCardType:[cardJSON[@"details"][@"cardType"] asString]]
                                        lastTwo:[cardJSON[@"details"][@"lastTwo"] asString]
+                                      lastFour:[cardJSON[@"details"][@"lastFour"] asString]
                                      isDefault:[cardJSON[@"default"] isTrue]
                                       cardJSON:cardJSON
                                authInsightJSON:authInsightJson];
 }
 
 + (instancetype)cardNonceWithGraphQLJSON:(BTJSON *)json {
-    NSString *lastFour = [json[@"creditCard"][@"last4"] asString];
+    NSString *lastFour = @"";
+    if ([json[@"creditCard"][@"last4"] asString]) {
+        lastFour = [json[@"creditCard"][@"last4"] asString];
+    }
     NSString *lastTwo = lastFour.length == 4 ? [lastFour substringFromIndex:2] : @"";
     NSString *description = lastTwo.length > 0 ? [NSString stringWithFormat:@"ending in %@", lastTwo] : @"";
     
@@ -110,10 +119,11 @@
         authInsightJson = json[@"authenticationInsight"];
     }
     
-    return [[[self class] alloc] initWithNonce:[json[@"token"] asString]
+    return [[self.class alloc] initWithNonce:[json[@"token"] asString]
                                    description:description
                                    cardNetwork:[self.class cardNetworkFromGatewayCardType:[json[@"creditCard"][@"brand"] asString]]
                                        lastTwo:lastTwo
+                                      lastFour:lastFour
                                      isDefault:NO
                                       cardJSON:json[@"creditCard"]
                                authInsightJSON:authInsightJson];