+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 
 * Add support for iOS 13 SceneDelegate to `BTAppSwitch`
+* Add `lastFour` property to `BTCardNonce`
 
 ## 4.32.1 (2020-02-21)